Backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart